+++ internal/dependencies/installer_command_other.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package dependencies
+
+import "os/exec"
+
+func configureInstallerCommand(_ *exec.Cmd) error { return nil }